@pellared 2023-12-06
Can you please run all benchmarks for the whole package? The changes should also affect `BenchmarkNew` and `BenchmarkParse`.